Move country tuples into the generic.countries submodule.
